Notion API: Move page discovery off the importer
Turning a search response into the tree the picker draws needs nothing
from the vault or the network, but it sat on the importer where no test
could reach it - so diagnosing a report of a missing page meant
reimplementing the filtering in a scratch script and trusting that the
copy agreed with the original.

Moved as it stands. The two filter phases, extractItemTitle,
extractParentId and buildTree are the same code with this. instead of a
receiver, and NotionParent and NotionTreeNode move with them.

Checked against a real workspace of 1123 search results: same 1064 items
and same 7 roots, with the same children under each.
